% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcbc-fonts.R
\name{lcbc_install_font}
\alias{lcbc_install_font}
\title{Install Avenir font for LCBC branding}
\usage{
lcbc_install_font(...)
}
\arguments{
\item{...}{arguments to \code{\link[extrafont]{font_import}}}
}
\description{
Install Avenir font for LCBC branding
}
\examples{
lcbc_install_font()
}
