% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcbc-logo.R
\name{lcbc_logo_list}
\alias{lcbc_logo_list}
\title{List available LCBC logo variations}
\usage{
lcbc_logo_list()
}
\value{
character vector of logo names
}
\description{
List available LCBC logo variations
}
\examples{
lcbc_logo_list()
}
