% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-paper-project.R
\name{create_paper_project}
\alias{create_paper_project}
\title{Initiate a paper project}
\usage{
create_paper_project(dir, title, edit = interactive())
}
\arguments{
\item{dir}{directory to create the project in}

\item{title}{Title of the project}

\item{edit}{Should files to edit be opened.}
}
\value{
no return
}
\description{
Initiate a paper project
}
